# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nombre de tu API",
        "version": "1.0.0",
        "description": "Descripción de tu API"
    },
    "servers": [{
        "url": "http://localhost:8080",
        "description": "Servidor de desarrollo"
    }],
    "paths": {
...
        "/api/series": {
            "get": {
                "tags": ["Series"],
                "summary": "Obtiene todas las series paginadas.",
                "responses": {
                    "200": {
                        "description": "Lista de todas las series."
                    },
                    "404": {
                        "description": "No se encontró"
                    }
                }
            },
            "/series/{id}": {
                "get": {
                    "tags": ["Series"],
                    "summary": "Obtiene una serie por su ID.",
                    "parameters": [{
                        "name": "id",
                        "in": "path",
                        "description": "ID de la serie a buscar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }],
                    "responses": {
                        "200": {
                            "description": "La serie encontrada."
                        },
                        "404": {
                            "description": "No se encontró"
                        }
                    }
                },
                "put": {
                    "tags": ["Series"],
                    "summary": "Actualiza una serie existente.",
                    "parameters": [{
                            "name": "id",
                            "in": "path",
                            "description": "ID de la serie a actualizar.",
                            "required": true,
                            "schema": {
                                "type": "integer"
                            }
                        },
                        {
                            "name": "img",
                            "in": "formData",
                            "description": "Imagen de la serie a subir (opcional).",
                            "required": false,
                            "schema": {
                                "type": "string",
                                "format": "binary"
                            }
                        },
                        {
                            "name": "name",
                            "in": "formData",
                            "description": "Nuevo nombre de la serie.",
                            "required": true,
                            "schema": {
                                "type": "string"
                            }
                        },
                        {
                            "name": "descr",
                            "in": "formData",
                            "description": "Nueva descripción de la serie.",
                            "required": true,
                            "schema": {
                                "type": "string"
                            }
                        },
                        {
                            "name": "search",
                            "in": "formData",
                            "description": "Nueva cadena de búsqueda de la serie.",
                            "required": true,
                            "schema": {
                                "type": "string"
                            }
                        },
                        {
                            "name": "categories",
                            "in": "formData",
                            "description": "Conjunto de identificadores de categorías asociadas a la serie.",
                            "required": true,
                            "schema": {
                                "type": "array",
                                "items": {
                                    "type": "integer"
                                }
                            }
                        }
                    ],
                    "responses": {
                        "201": {
                            "description": "La serie actualizada."
                        },
                        "409": {
                            "description": "Nombre duplicado"
                        },
                        "404": {
                            "description": "No se encontró la serie"
                        },
                        "422": {
                            "description": "Fallo en la validación, devuelve Campo:Error"
                        }
                    }
                },
                "delete": {
                    "tags": ["Series"],
                    "summary": "Elimina una serie por su ID.",
                    "parameters": [{
                        "name": "id",
                        "in": "path",
                        "description": "ID de la serie a eliminar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }],
                    "responses": {
                        "204": {
                            "description": "Operación exitosa."
                        }
                    }
                }
            },
            "post": {
                "tags": ["Series"],
                "summary": "Crea una nueva serie con imagen y categorías asociadas.",
                "parameters": [{
                        "name": "img",
                        "in": "formData",
                        "description": "El archivo de imagen de la serie.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "binary"
                        }
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "El nombre de la serie.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "descr",
                        "in": "formData",
                        "description": "La descripción de la serie.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "formData",
                        "description": "El término de búsqueda de la serie.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "categories",
                        "in": "formData",
                        "description": "El conjunto de identificadores de categorías asociadas.",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "La serie actualizada."
                    },
                    "409": {
                        "description": "Nombre duplicado"
                    },
                    "422": {
                        "description": "Fallo en la validación, devuelve Campo:Error"
                    }
                }
            }
        },
...
        "/api/users": {
            "get": {
                "tags": ["Users"],
                "summary": "Obtiene todos los usuarios.",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios."
                    },
                    "404": {
                        "description": "No se encontró"
                    }
                }
            },
            "/{id}": {
                "get": {
                    "tags": ["Users"],
                    "summary": "Obtiene un usuario por su ID.",
                    "parameters": [{
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario a obtener.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }],
                    "responses": {
                        "200": {
                            "description": "El usuario correspondiente."
                        },
                        "404": {
                            "description": "No se encontró"
                        }
                    }
                },
                "put": {
                    "tags": ["Users"],
                    "summary": "Actualiza un usuario por su ID.",
                    "parameters": [{
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario a actualizar.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }],
                    "requestBody": {
                        "description": "Datos del usuario a actualizar.",
                        "required": true,
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetUserDTO"
                                }
                            }
                        }
                    },
                    "responses": {
                        "200": {
                            "description": "El usuario actualizado."
                        },
                        "404": {
                            "description": "No se encontró"
                        },
                        "409": {
                            "description": "Email duplicado"
                        },
                        "422": {
                            "description": "Error de validación, devuelve objeto Campo:error"
                        }

                    }
                },
                "delete": {
                    "tags": ["Users"],
                    "summary": "Elimina un usuario por su ID.",
                    "parameters": [{
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario a eliminar.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }],
                    "responses": {
                        "204": {
                            "description": "No Content."
                        },
                        "404": {
                            "description": "No se encontró"
                        }
                    }
                }
            },
            "post": {
                "tags": ["Users"],
                "summary": "Agrega un nuevo usuario.",
                "requestBody": {
                    "description": "Datos del usuario a crear.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostUserDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "El usuario creado."
                    },

                    "409": {
                        "description": "Email duplicado"
                    },
                    "422": {
                        "description": "Error de validación, devuelve objeto Campo:error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PostUserDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "El nombre del usuario.",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "description": "El correo electrónico del usuario.",
                        "example": "johndoe@example.com"
                    },
                    "password": {
                        "type": "string",
                        "description": "La contraseña del usuario.",
                        "example": "password123"
                    },
                    "phone": {
                        "type": "string",
                        "description": "El número de teléfono del usuario.",
                        "example": "+1234567890"
                    },
                    "rol": {
                        "type": "integer",
                        "description": "El rol del usuario.",
                        "example": 0
                    }
                }
            }
...
            "GetUserDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "El identificador único del usuario."
                    },
                    "name": {
                        "type": "string",
                        "description": "El nombre del usuario.",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "description": "La dirección de correo electrónico del usuario.",
                        "example": "john@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "description": "El número de teléfono del usuario.",
                        "example": "123456789"
                    },
                    "rol": {
                        "type": "integer",
                        "description": "El rol del usuario."
                    },
                    "createdAt": {
                        "type": "string",
                        "description": "La fecha y hora en que se creó el usuario.",
                        "example": "2024-04-08"
                    },
                    "dischargeDate": {
                        "type": "string",
                        "description": "La fecha de baja del usuario.",
                        "nullable": true
                    },
                    "imgPath": {
                        "type": "string",
                        "description": "La ruta de la imagen del usuario.",
                        "example": "default.png"
                    }
                },
                "required": ["id", "name", "email", "phone"],
                "title": "GetUserDTO"
            }
...
        }
    }
}